rwkv_peft/requirements.txt
pytorch-lightning==1.9.5
bitsandbytes
deepspeed
einops
triton==2.2.0